+++ src/license_expression/__init__.py
@@ -1796,6 +1796,14 @@ def combine_expressions(
         'gpl AND mit AND apache'
         >>> str(combine_expressions(('gpl', 'mit', 'apache',), relation='OR'))
         'gpl OR mit OR apache'
+        >>> str(combine_expressions(('gpl', 'mit', 'mit',)))
+        'gpl AND mit'
+        >>> str(combine_expressions(('mit WITH foo', 'gpl', 'mit',)))
+        'mit WITH foo AND gpl AND mit'
+        >>> str(combine_expressions(('gpl', 'mit', 'mit',), relation='OR', unique=False))
+        'gpl OR mit OR mit'
+        >>> str(combine_expressions(('mit', 'gpl', 'mit',)))
+        'mit AND gpl'
     """
     if not expressions:
         return
@@ -1805,11 +1813,16 @@ def combine_expressions(
             f'expressions should be a list or tuple and not: {type(expressions)}'
         )
 
-    # only del with LicenseExpression objects
+    if not relation or relation.upper() not in ('AND', 'OR',):
+        raise TypeError(f'relation should be one of AND, OR and not: {relation}')
+
+
+    # only deal with LicenseExpression objects
     expressions = [licensing.parse(le, simple=True) for le in expressions]
 
-    # Remove duplicate expressions
     if unique:
+        # Remove duplicate element in the expressions list
+        # and preserve original order
         expressions = list({str(x): x for x in expressions}.values())
 
     if len(expressions) == 1: